requirements.txt
dash
plotly
shimmy>=2.0
ollama
asyncio
openai
pandas
requests
aiofiles
stable-baselines3
torch
gym
deap
numpy
configparser
python-dotenv
stable-baselines3[extra]
dash-bootstrap-components
fastapi
uvicorn
